requirements.txt
boto3>=1.34.0
langchain-aws>=0.2.0
langchain-openai>=0.3.0
langgraph>=0.2.0
pinecone>=6.0.0
pypdf>=5.0.0
python-dotenv>=1.0.1
